feat(TTB-043): Moved header files to common directory
